# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Resource Booking System API",
        "description": "API documentation for the Resource Booking Management System",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/resources": {
            "get": {
                "tags": [
                    "Resources"
                ],
                "summary": "List all resources",
                "operationId": "7633cd726de24dafbff1c0559ade67e2",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filter resources by type",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "active_only",
                        "in": "query",
                        "description": "Filter only active resources",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of resources",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Resource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Resources"
                ],
                "summary": "Create a new resource",
                "operationId": "ae1560ca6061943b6da1b27ed6e18059",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreResourceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Resource created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Resource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/resources/{id}": {
            "get": {
                "tags": [
                    "Resources"
                ],
                "summary": "Get a specific resource",
                "operationId": "1c48e8d46ffeee459fdeccd77f879645",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resource details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Resource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Resources"
                ],
                "summary": "Update a resource",
                "operationId": "0b3ae63a2994a393c57d2d5ba53e0913",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateResourceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resource updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Resource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Resources"
                ],
                "summary": "Delete a resource",
                "operationId": "edabc99f680998ad2bda300c06eb0bab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Resource deleted successfully"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "409": {
                        "description": "Resource cannot be deleted due to active bookings"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Resource": {
                "title": "Resource",
                "description": "Resource model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID of the resource",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name of the resource",
                        "type": "string",
                        "example": "Meeting Room A"
                    },
                    "type": {
                        "title": "Type",
                        "description": "Type of the resource",
                        "type": "string",
                        "example": "room"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description of the resource",
                        "type": "string",
                        "example": "A spacious meeting room with projector"
                    },
                    "is_active": {
                        "title": "Is Active",
                        "description": "Whether the resource is active",
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2025-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2025-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Resource"
                }
            },
...
            "StoreResourceRequest": {
                "title": "Store Resource Request",
                "description": "Store Resource request body data",
                "required": [
                    "name",
                    "type"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the resource",
                        "type": "string",
                        "example": "Meeting Room A"
                    },
                    "type": {
                        "title": "type",
                        "description": "Type of the resource",
                        "type": "string",
                        "example": "room"
                    },
                    "description": {
                        "title": "description",
                        "description": "Description of the resource",
                        "type": "string",
                        "example": "A spacious meeting room with projector"
                    },
                    "is_active": {
                        "title": "is_active",
                        "description": "Whether the resource is active",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "UpdateResourceRequest": {
                "title": "Update Resource Request",
                "description": "Update Resource request body data",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the resource",
                        "type": "string",
                        "example": "Meeting Room A"
                    },
                    "type": {
                        "title": "type",
                        "description": "Type of the resource",
                        "type": "string",
                        "example": "room"
                    },
                    "description": {
                        "title": "description",
                        "description": "Description of the resource",
                        "type": "string",
                        "example": "A spacious meeting room with projector"
                    },
                    "is_active": {
                        "title": "is_active",
                        "description": "Whether the resource is active",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}